requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

avalanche-lib
pandas
jupyter
matplotlib
pytorchcv==0.0.67

pytest